Add JUnit tests for listEmployeeInLeaveCommand
